+++ e310x-hal/src/asynch/embassy.rs
@@ -1,0 +1,3 @@
+//! Module with all embassy-related code
+
+pub mod time_driver;